requirements.txt
# Core LLM providers
anthropic>=0.18.0
openai>=1.12.0

# GitHub API
PyGithub>=2.1.1

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.6.0

# HTTP requests
requests>=2.31.0

# Code quality (optional)
pylint>=3.0.0
black>=24.1.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
